refactor(nats): #1592 split render_event_codec.py v2 registry into registry module
Extract v2 registry entries (13 CodecBranch definitions) from
render_event_codec.py into new render_event_codec_registry.py.

- render_event_codec_registry.py: CodecBranch + _make_std_encode/decode +
  build_codec_registry() with all 13 event-type branches
- render_event_codec.py: retains NatsRenderEventCodec class with
  encode/decode/is_terminal + synthetic terminal constants
  Shrinks 388→191 lines (target <300 achieved)
- render_event_codec.py no longer needs json/dataclasses/Callable imports
- Removed file_exemptions.txt entry for render_event_codec.py
- Updated folder_exemptions.txt src/lyra/nats 16→20
