+++ tools/folder_exemptions.txt
@@ -11,4 +11,4 @@ src/lyra/infrastructure/stores  # 16 files — DEBT:folder-exemptions — #935 A
 src/lyra/core/agent         # 14 files — DEBT:folder-exemptions — #1414 bot_models.py + bot_schema.py added; refactor to core/bots/ or similar in Phase 7
 src/lyra/core/cli           # 14 files — DEBT:folder-exemptions — #957 cli_pool_entry.py extracted to break _ProcessEntry circular import
 src/lyra/core/messaging     # 13 files — DEBT:folder-exemptions — #1016 WorkerError: error_extractor.py + metrics.py added; #1031 tool_recap_format.py + tool_display_config.py added
-src/lyra/nats  # 19 files — DEBT:folder-exemptions — #1278 codec split from clients (T11/T12 added nats_{tts,stt,image}_codec.py); _worker_client_base.py + nats_llm_client.py deleted (net +1 vs 15 pre-#1278 baseline: -2 deletions + 3 codec files); #1589 extracted audio_publish.py + keepalive.py + stream_error.py from nats_channel_proxy.py
+src/lyra/nats  # 20 files — DEBT:folder-exemptions — #1278 codec split from clients (T11/T12 added nats_{tts,stt,image}_codec.py); _worker_client_base.py + nats_llm_client.py deleted (net +1 vs 15 pre-#1278 baseline: -2 deletions + 3 codec files); #1589 extracted audio_publish.py + keepalive.py + stream_error.py from nats_channel_proxy.py; #1592 extracted render_event_codec_registry.py from render_event_codec.py
